requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
gunicorn==20.1.0
python-dotenv==1.0.0
Werkzeug>=2.2.2